requirements.txt
pandas
matplotlib
seaborn
fastapi
uvicorn[standard]
firebase-admin
python-dotenv